Adding hash test
